# maquinaria_laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Maquinaria",
        "description": "Documentación de la API de Mantenimientos y Maquinarias",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/tipos-maquinaria": {
            "get": {
                "tags": [
                    "Tipos de Maquinaria"
                ],
                "summary": "Listar todos los tipos de maquinaria",
                "operationId": "5a603a21ad7e77667670201e6cd9be4a",
                "responses": {
                    "200": {
                        "description": "Listado de tipos de maquinaria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TipoMaquinaria"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos de Maquinaria"
                ],
                "summary": "Crear un nuevo tipo de maquinaria",
                "operationId": "b9972420abd6ea3bd4ca6da14f2c7d2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Excavadora"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Maquinaria pesada usada para excavación"
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de maquinaria creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de maquinaria creado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TipoMaquinaria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tipos-maquinaria/{id}": {
            "get": {
                "tags": [
                    "Tipos de Maquinaria"
                ],
                "summary": "Obtener un tipo de maquinaria por ID",
                "operationId": "b50a0a6cd7510b7474ef595817a308c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de maquinaria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de maquinaria encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TipoMaquinaria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de maquinaria no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipos de Maquinaria"
                ],
                "summary": "Actualizar un tipo de maquinaria",
                "operationId": "fe54e1a1eff2e3bc00327aca56efc351",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de maquinaria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Retroexcavadora"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Equipo usado para remover tierra"
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de maquinaria actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de maquinaria actualizado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TipoMaquinaria"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipos de Maquinaria"
                ],
                "summary": "Eliminar un tipo de maquinaria",
                "operationId": "f085956acdab1020c9c7c22fa622b983",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de maquinaria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de maquinaria eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de maquinaria eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de maquinaria no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TipoMaquinaria": {
                "title": "Tipo de Maquinaria",
                "required": [
                    "nombre",
                    "categoria_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Excavadora"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Equipo de construcción usado para excavar"
                    },
                    "categoria_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-30T14:48:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-30T14:48:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}